Frontend connects to backend
Connected the frontend to the backend now.
Additionally did some folder renaming for clarity.
Installed the required cors package to allow the backend to act as an api.
Ran the first migration and initialized the CustomUser model.
Lastly, set up the index.html js file to handle the frontend logic.