Create Assignment Endpoint Configured + Refactoring
The create assignment endpoint at POST  /api/assignments/
is configured and working via POSTMAN test. Additionally
did some refactoring for clariy and code reusability in the other files,
check the changes. IMPORTANT CHANGE IN models.py, using the modern
mapped and mapped column attributes now in the database models definition,
this is so mypy type checking works properly. No need for a migration, for
this change.