DELETE ENDPOINT CONFIGURED -> READ BODY
delete endpoint is up working and tested. Works well.
Additionally migrated the database to add on delete cascading.
Now if a user is deleted all assignments are deleted, if an assignment is deleted,
all related stats are deleted. Next is to refactor some routing endpoints to include
the user id token.