UPDATES -> READ BODY
FRONTEND: Added an audio cue that is played from the users browser whenever
an assignment is completed, whether its a manual completion
or automatic one.

BACKEND: Installed typer to create a CLI command.
It is a cleanup command that deletes all users in the database that are expired at the time
that the command is ran. IN production it will be used via a cronjob.
Additionally, fixed the logging config. Was not working since the handlers were outputting the logs to the wrong
files. It has been fixed. No ruff linting or formatting issues, mypy has no warnings.
Will use this timer myself for a few hours today, then ill begin the deployment.