Fixed frontend bug
Frontend was manually updating the timer via +1 intervals.
Issue is, when the tab is not the main focus in the browser,
the intervals throttle and stop. Switched to this, now using date.now
and start time to ensure that the time intervals are always correct when the timer
is not paused.