Frontend Bug Fixes
Fixed bugs in the frontend

Most prominent bug being the calculation of elapsed time with pauses.
Before it would do datetime.now - start time, but this would be inaccurate
as it wouldn't consider the pause time, added a new field pausetime, that
sets start time to the pause time so now even with pauses,
the elapsed time is always accurate.